pyproject.toml
[project]
name = "scout-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "python-dotenv>=1.1.0",
    "yt-dlp>=2025.5.22",
]

[tool.setuptools]
packages = ["src"]
